+++ .github/workflows/manual-trigger.yml
@@ -2,7 +2,8 @@
 
 name: Workflow - Run manual trigger
 
-# Controls when the action will run. Workflow runs when manually triggered
+# Controls when the action will run. Workflow runs when manually triggered using the UI
+# or API.
 on:
   workflow_dispatch:
     # Inputs the workflow accepts.
@@ -17,8 +18,7 @@ on:
         # The data type of the input
         type: string
 env:
-  TARGET_ENV: ${{ inputs.name }}
-  GH_TOKEN: ${{secrets.GH_API_TOKEN}}
+  TARGET_ENV: ${{ inputs.name }} 
   OC_CLUSTER: ${{ vars.OPENSHIFT_CLUSTER }}
   OC_REGISTRY: ${{ vars.OPENSHIFT_REGISTRY }}
   OC_AUTH_TOKEN: ${{ secrets.OPENSHIFT_TOKEN }}
@@ -30,95 +30,55 @@ env:
   UGM_BUILD_VERSION: ${{vars.UGM_BUILD_VERSION}}
   UGM_BUILD_REVISION: ${{vars.UGM_BUILD_REVISION}}
 
-  UGM_RELEASE_PREFIX: Unity_Grant_Manager_v
-  UGM_RELEASE_MESSAGE: "Manual deployment"
-
+# A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
-  Setup:
+  setup:
+    # The type of runner that the job will run on
     runs-on: ubuntu-latest
+    # Steps represent a sequence of tasks that will be executed as part of the job
     environment: ${{ inputs.name }}
     steps:
+    # Runs commands using the runners shell
     - name: Get variables
       run: |
-        echo "Target: $TARGET_ENV"
-        echo "BaseRef: $GITHUB_REF_NAME"
+        echo "target=$TARGET_ENV "
+        echo "base=$GITHUB_REF_NAME"
         echo "Environment: $TARGET_ENV OC_TARGET_PROJECT=$OC_TARGET_PROJECT"
         echo "Environment: $TARGET_ENV JFROG_REPO_PATH=$JFROG_REPO_PATH"
         echo "..."
         env | sort
-    - name: Get current date
-      id: date_selector
-      run: echo "DATE=$(date +'%B %e, %Y')" >> $GITHUB_OUTPUT
-    outputs:
-      DATE: ${{steps.date_selector.outputs.DATE}}
-  Branch:
-    needs: [Setup]
-    runs-on: ubuntu-latest
-    environment: ${{ inputs.name }}
-    steps:
-    - name: Checkout repository
-      uses: actions/checkout@v4
-      with:
-          fetch-depth: '1'
-    - name: Get short commitId
-      id: get_commit
-      run:  |
-        echo "SHA_SHORT=$(git rev-parse --short HEAD)" >> $GITHUB_OUTPUT
-    - name: List merged branch
-      id: get_branch
-      run: |
-        MERGED_BRANCH=$(git log --merges --pretty=format:"%s" -1 | grep -oE 'from bcgov/[^ ]+' | sed 's/from bcgov\///')
-        if [ -z "$MERGED_BRANCH" ]; then
-          echo "Direct push no pull request from bcgov/branch"
-          echo "MERGED_BRANCH=push" >> $GITHUB_OUTPUT
-        else
-          echo "Merged branch: $MERGED_BRANCH"
-          echo "MERGED_BRANCH=$MERGED_BRANCH" >> $GITHUB_OUTPUT
-        fi
-    outputs:
-      SHA_SHORT: ${{steps.get_commit.outputs.SHA_SHORT}}
-      MERGED_BRANCH: ${{steps.get_branch.outputs.MERGED_BRANCH}}
-  PushVariables:
-    needs: [Setup,Branch]
-    runs-on: ubuntu-latest
-    environment: ${{ inputs.name }}
-    steps:
-    - name: Checkout repository
-      uses: actions/checkout@v4
-      with:
-          fetch-depth: '1'
-    - name: Set repository version variables
-      id: set_version
-      run: |
-        # gh variable set UGM_BUILD_REVISION --env ${{env.TARGET_ENV}} --body "${{needs.Branch.outputs.SHA_SHORT}}"
-        echo "UGM_BUILD_REVISION=${{needs.Branch.outputs.SHA_SHORT}}" >> $GITHUB_ENV
-    - name: Get repository version variables
-      id: get_version
-      run: |
-        gh variable list --env ${{env.TARGET_ENV}}
-        echo "..."
-        echo "buildArgs --env ${{env.TARGET_ENV}} UNITY_BUILD_VERSION: ${{env.UGM_BUILD_VERSION}}, UNITY_BUILD_REVISION: ${{env.UGM_BUILD_REVISION}}, Merged Branch: ${{needs.Branch.outputs.MERGED_BRANCH}}"
-  Build:
-    needs: [Setup,Branch,PushVariables]
+  build:
     runs-on: ubuntu-latest
     environment: ${{ inputs.name }}
     steps:
     - uses: actions/checkout@v4
+    - name: Build dotnetapp docker images
+      run: docker compose build
+      working-directory: ./applications/dotnetapp
+    - name: Build aspnetapp docker images
+      run: docker compose build
+      working-directory: ./applications/aspnetapp
+    - name: Get BuildId
+      id: commit
+      run: echo "sha_short=$(git rev-parse --short HEAD)" >> $GITHUB_OUTPUT
+    - name: Set buildArgs
+      run:  |
+        echo "buildArgs UNITY_BUILD_VERSION: ${{env.UGM_BUILD_VERSION}}, UNITY_BUILD_REVISION: ${{env.UGM_BUILD_REVISION}}"
     - name: Build Docker images
       run: |
         rm -f ./docker-compose.override.yml
-        echo "buildArgs UNITY_BUILD_VERSION: ${{env.UGM_BUILD_VERSION}}, UNITY_BUILD_REVISION: ${{env.UGM_BUILD_REVISION}}"
         docker compose build --build-arg UNITY_BUILD_VERSION=${{env.UGM_BUILD_VERSION}} --build-arg UNITY_BUILD_REVISION=${{env.UGM_BUILD_REVISION}}
       working-directory: ./applications/Unity.GrantManager
+
     - name: Connect to JFrog Artifactory non-interactive login using --password-stdin
       run: |
         echo "$JFROG_PASSWORD" | docker login -u "$JFROG_USERNAME" --password-stdin $JFROG_SERVICE
     - name: Push application images to Artifactory container registry
       run:  |
         docker tag unity-grantmanager-dbmigrator $JFROG_SERVICE/$JFROG_REPO_PATH/unity-grantmanager-dbmigrator
-        # docker push $JFROG_SERVICE/$JFROG_REPO_PATH/unity-grantmanager-dbmigrator
+        docker push $JFROG_SERVICE/$JFROG_REPO_PATH/unity-grantmanager-dbmigrator
         docker tag unity-grantmanager-web $JFROG_SERVICE/$JFROG_REPO_PATH/unity-grantmanager-web
-        # docker push $JFROG_SERVICE/$JFROG_REPO_PATH/unity-grantmanager-web
+        docker push $JFROG_SERVICE/$JFROG_REPO_PATH/unity-grantmanager-web
     - name: Disconnect docker from JFrog Artifactory
       run: |
         docker logout
@@ -130,9 +90,9 @@ jobs:
     - name: Push application images to OpenShift container registry
       run:  |
         docker tag unity-grantmanager-dbmigrator $OC_REGISTRY/$OC_TARGET_PROJECT/unity-grantmanager-dbmigrator
-        # docker push $OC_REGISTRY/$OC_TARGET_PROJECT/unity-grantmanager-dbmigrator
+        docker push $OC_REGISTRY/$OC_TARGET_PROJECT/unity-grantmanager-dbmigrator
         docker tag unity-grantmanager-web $OC_REGISTRY/$OC_TARGET_PROJECT/unity-grantmanager-web
-        # docker push $OC_REGISTRY/$OC_TARGET_PROJECT/unity-grantmanager-web
+        docker push $OC_REGISTRY/$OC_TARGET_PROJECT/unity-grantmanager-web
     - name: Disconnect docker from OpenShift container registry
       run: |
         docker logout
