+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/ApplicationForms/ApplicationFormDto.cs
@@ -20,7 +20,6 @@ namespace Unity.GrantManager.ApplicationForms
         public string? ConnectionHttpStatus { get; set; }
         public DateTime? AttemptedConnectionDate { get; set; }
         public bool Payable { get; set; }
-        public bool RenderFormIoToHtml { get; set; }
         public Guid? ScoresheetId { get; set; }
         public Guid? TenantId { get; set; }
     }
