+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/Intakes/IApplicantService.cs
@@ -2,7 +2,7 @@
 
 namespace Unity.GrantManager.Intakes
 {
-    public interface IApplicantLookupService
+    public interface IApplicantService
     {
         Task<string> ApplicantLookupByApplicantId(string? unityApplicantId);
     }
