+++ applications/Unity.GrantManager/src/Unity.GrantManager.HttpApi/Controllers/ApplicantLookupController.cs
@@ -7,7 +7,6 @@ using Unity.GrantManager.Data;
 using Unity.GrantManager.Controllers.Auth.FormSubmission;
 using Unity.GrantManager.Intakes;
 
-
 namespace Unity.GrantManager.Controllers
 {
     [ApiController]
@@ -15,42 +14,31 @@ namespace Unity.GrantManager.Controllers
     [AllowAnonymous]
     public class ApplicantLookupController(
                                 ITenantRepository tenantRepository,
-                                IApplicantLookupService applicantService) : AbpControllerBase
+                                IApplicantService applicantService) : AbpControllerBase
     {
 #pragma warning disable ASP0018
         // Needed for the Database Context
         [HttpGet("{__tenant}")]
 #pragma warning restore ASP0018
         [ServiceFilter(typeof(FormsApiTokenAuthFilter))]
-        public async Task<IActionResult> GetApplicantAsync([FromQuery] ApplicantLookup applicantLookup)
+        public async Task<dynamic> GetApplicantAsync([FromQuery] ApplicantLookup applicantLookup)
         {
+
             if (applicantLookup.UnityApplicantId == null)
             {
-                return NotFound("Applicant Not Found");
+                return "Applicant NotFound";
             }
 
-            // Handle tenant context
             if (CurrentTenant.Id == null)
             {
                 var defaultTenant = await tenantRepository.FindByNameAsync(GrantManagerConsts.NormalizedDefaultTenantName);
                 using (CurrentTenant.Change(defaultTenant.Id, defaultTenant.Name))
                 {
-                    return await GetApplicantContent(applicantLookup.UnityApplicantId);
+                    var currentResult = await applicantService.ApplicantLookupByApplicantId(applicantLookup.UnityApplicantId);
+                    return Content(currentResult, "application/json");
                 }
-            }
-
-            return await GetApplicantContent(applicantLookup.UnityApplicantId);
-        }
-
-        private async Task<IActionResult> GetApplicantContent(string unityApplicantId)
-        {
-            var result = await applicantService.ApplicantLookupByApplicantId(unityApplicantId);
-            
-            if (result == null)
-            {
-                return NotFound("Applicant Not Found");
-            }
-
+            } 
+            var result = await applicantService.ApplicantLookupByApplicantId(applicantLookup.UnityApplicantId);
             return Content(result, "application/json");
         }
     }
