+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Pages/ApplicationForms/ViewModels/CreateUpdateApplicationFormViewModel.cs
@@ -15,7 +15,7 @@ namespace Unity.GrantManager.Web.Pages.ApplicationForms.ViewModels
         [Required]
         public List<SelectListItem> IntakesList { get; set; } = new List<SelectListItem>();
 
-        [ReadOnlyInput]
+        [DisabledInput]
         [DisplayName("Common:Name")]
         public string ApplicationFormName { get; set; } = string.Empty;
 
@@ -35,8 +35,5 @@ namespace Unity.GrantManager.Web.Pages.ApplicationForms.ViewModels
 
         [DisplayName("ApplicationForms:Payable")]
         public bool Payable { get; set; }
-
-        [DisplayName("ApplicationForms:RenderFormIoToHtml")]
-        public bool RenderFormIoToHtml { get; set; }
     }
 }
