+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Pages/GrantApplications/Details.js
@@ -1,6 +1,5 @@
 $(function () {
     let selectedReviewDetails = null;
-    let renderFormIoToHtml = document.getElementById('RenderFormIoToHtml').value;
     let hasRenderedHtml = document.getElementById('HasRenderedHTML').value;
     abp.localization.getResource('GrantManager');
 
@@ -13,7 +12,7 @@ $(function () {
     initializeDetailsPage();
 
     function renderSubmission() {
-        if (renderFormIoToHtml == "False" || hasRenderedHtml == "False") {
+        if (hasRenderedHtml == "False") {
             getSubmission();
         } else {
             $('.spinner-grow').hide();
@@ -38,7 +37,7 @@ $(function () {
             let submissionData = JSON.parse(submissionString);
             Formio.icons = 'fontawesome';
 
-            await Formio.createForm(
+            Formio.createForm(
                 document.getElementById('formio'),
                 submissionData.version.schema,
                 {
@@ -47,34 +46,29 @@ $(function () {
                     flatten: true,
                 }
             ).then(function (form) {
-                handleForm(form, submissionData.submission.submission);
-            });
+                // Set Example Submission Object
+                form.submission = submissionData.submission.submission;
+                form.resetValue();
+                form.refresh();
+                form.on('render', function () {
+                    addEventListeners();
+                });
 
+                waitFor(_ => isFormChanging(form))
+                    .then(_ => 
+                        storeRenderedHtml()
+                    );
+                });
         } catch (error) {
             console.error(error);
         }
     }
-
-    function handleForm(form, submission) {
-        form.submission = submission;
-        form.resetValue();
-        form.refresh();
-        form.on('render', addEventListeners);
-
-        waitFor(() => isFormChanging(form)).then(() => {
-            setTimeout(storeRenderedHtml, 2000);
-        });
-    }
-
 
     function isFormChanging(form) {
         return form.changing === false;
     }
 
     async function storeRenderedHtml() {
-        if (renderFormIoToHtml == "False") {
-            return;
-        }
         let innerHTML = document.getElementById('formio').innerHTML;
         let submissionId = document.getElementById('ApplicationFormSubmissionId').value;
         $.ajax(
@@ -97,10 +91,10 @@ $(function () {
     function addEventListeners() {
         const cardHeaders = getCardHeaders();
         const cardBodies = getCardBodies();
-
+        
         // Collapse all card bodies initially
         hideAllCardBodies(cardBodies);
-
+        
         // Add event listeners to headers
         cardHeaders.forEach(header => {
             header.addEventListener('click', () => onCardHeaderClick(header, cardHeaders));
@@ -287,7 +281,7 @@ $(function () {
 
     $('#printPdf').click(function () {
         let submissionId = getSubmissionId();
-
+    
         // Fetch submission data
         fetchSubmissionData(submissionId)
             .done(function (result) {
@@ -297,19 +291,19 @@ $(function () {
                 console.error('Error fetching submission data:', error);
             });
     });
-
+    
     // Get submission ID from the input field
     function getSubmissionId() {
         return document.getElementById('ChefsSubmissionId').value;
     }
-
+    
     // Fetch the submission data
     function fetchSubmissionData(submissionId) {
         return unity.grantManager.intakes.submission.getSubmission(submissionId);
     }
-
+    
     // Handle the submission result
-    function openDataInNewTab(data, submissionId) {
+    function openDataInNewTab(data, submissionId) {        
         let newTab = window.open('', '_blank');
         let newDiv = $('<div>');
 
@@ -321,7 +315,7 @@ $(function () {
 
         // Store the outer HTML of the new div in divToStore
         let divToStore = newDiv.prop('outerHTML');
-
+        
         newTab.document.write('<html><head><title>Print</title>');
         newTab.document.write('<script src="/libs/jquery/jquery.js"></script>');
         newTab.document.write('<script src="/libs/formiojs/formio.form.js"></script>');
@@ -391,7 +385,7 @@ $(function () {
     const widgetCallback = function (mutationsList, observer) {
         for (const mutation of mutationsList) {
             if (mutation.type === 'childList') {
-                initCustomFieldCurrencies();
+                initCustomFieldCurrencies(); 
                 break;
             }
         }
@@ -402,10 +396,10 @@ $(function () {
 
     PubSub.subscribe(
         'application_status_changed',
-        (msg, data) => {
+        (msg, data) => {            
             applicationBreadcrumbWidgetManager.refresh();
             applicationStatusWidgetManager.refresh();
-            assessmentResultWidgetManager.refresh();
+            assessmentResultWidgetManager.refresh();            
         }
     );
 
@@ -415,7 +409,7 @@ $(function () {
             decimal: '.'
         }).maskMoney('mask');
     }
-
+    
     PubSub.subscribe('application_assessment_results_saved',
         (msg, data) => {
             assessmentResultWidgetManager.refresh();
@@ -432,7 +426,7 @@ $(function () {
             }
         }
     });
-
+    
     const summaryWidgetTargetNode = document.querySelector('#' + summaryWidgetDiv);
     const summaryWidgetObserver = new MutationObserver(widgetCallback);
     summaryWidgetObserver.observe(summaryWidgetTargetNode, widgetConfig);
@@ -488,7 +482,7 @@ $(function () {
         let applicationId = decodeURIComponent($("#DetailsViewApplicationId").val());
         let formData = $(`#${formDataName}`).serializeArray();
         let customFormObj = {};
-        let formVersionId = $("#ApplicationFormVersionId").val();
+        let formVersionId = $("#ApplicationFormVersionId").val();        
 
         $.each(formData, function (_, input) {
             customFormObj[input.name] = input.value;
@@ -503,11 +497,11 @@ $(function () {
 
     PubSub.subscribe(
         'fields_tab',
-        (_, data) => {
+        (_, data) => {          
             let formDataName = data.worksheet + '_form';
-            let formValid = $(`form#${formDataName}`).valid();
+            let formValid = $(`form#${formDataName}`).valid();               
             let saveBtn = $(`#save_${data.worksheet}_btn`);
-            if (formValid && !formHasInvalidCurrencyCustomFields(`${formDataName}`)) {
+            if (formValid && !formHasInvalidCurrencyCustomFields(`${formDataName}`)) {                
                 saveBtn.prop('disabled', false);
             } else {
                 saveBtn.prop('disabled', true);
