+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/appsettings.Development.json
@@ -54,8 +54,7 @@
           "rollingInterval": "Day",
           "rollOnFileSizeLimit": true,
           "retainedFileCountLimit": 31,
-          "formatter": "Serilog.Formatting.Json.JsonFormatter, Serilog",
-          "shared": true
+          "formatter": "Serilog.Formatting.Json.JsonFormatter, Serilog"
         }
       }
     ],
