+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/appsettings.json
@@ -131,8 +131,7 @@
           "path": "logs/log.txt",
           "rollingInterval": "Day",
           "rollOnFileSizeLimit": true,
-          "formatter": "Serilog.Formatting.Json.JsonFormatter, Serilog",
-          "shared": true
+          "formatter": "Serilog.Formatting.Json.JsonFormatter, Serilog"
         }
       }
     ],
